Fixes and improvements

+++ src/main/java/cn/nukkit/network/protocol/ContainerClosePacket.java
@@ -18,7 +18,7 @@ public class ContainerClosePacket extends DataPacket {
 
     @Override
     public void decode() {
-        this.windowId = this.getByte();
+        this.windowId = (byte) this.getByte();
     }
 
     @Override
